+++ src/lib/st-breadcrumbs/st-breadcrumbs.component.spec.ts
@@ -8,6 +8,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
+import { SimpleChange } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { Http } from '@angular/http';
 import { RouterTestingModule } from '@angular/router/testing';
@@ -15,19 +16,13 @@ import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { StBreadCrumbsComponent } from './st-breadcrumbs.component';
 
-describe('[BreadCrumbsComponent]', () => {
+describe('StBreadCrumbsComponent', () => {
    let component: StBreadCrumbsComponent;
    let fixture: ComponentFixture<StBreadCrumbsComponent>;
 
-   let windowRefServiceMock: any;
-   let menuMock1 = ['section1', 'section2', 'section3', 'section4', 'section5'];
-   let menuMock2 = [
-      'section6',
-      'section7',
-      'section8',
-      'section9',
-      'section10'
-   ];
+   const menuMock: string[] = ['section0', 'section1', 'section2', 'section3', 'section4', 'section5', 'section6'];
+   const expectedIndexShowing5: number[] = [0, -1, 3, 4, 5, 6];
+   const expectedIndexShowing10: number[] = [0, 1, 2, 3, 4, 5, 6];
 
    beforeEach(
       async(() => {
@@ -39,51 +34,74 @@ describe('[BreadCrumbsComponent]', () => {
    );
 
    beforeEach(() => {
-      windowRefServiceMock = {};
-      windowRefServiceMock.nativeWindow = window;
       fixture = TestBed.createComponent(StBreadCrumbsComponent);
       component = fixture.componentInstance;
+      component.options = menuMock;
    });
 
-   describe('if initialized with 6 elements or less', () => {
-      beforeEach(() => {
-         component.options = menuMock1;
-      });
+   it('Should be initialized with elements to show set to 10 options', () => {
+      component.elementsToShow = 10;
+      fixture.detectChanges();
 
-      it('All elements will be shown', () => {
-         expect(component.generateCrumbs()).toEqual(menuMock1);
-      });
-
-      it('And user clicks on an element, component emits the element position', () => {
-         spyOn(component.select, 'emit');
-         component.onSelect(2);
-         expect(component.select.emit).toHaveBeenCalledWith(2);
-      });
-
-      it('If user clicks on the active element, nothing happens', () => {
-         component.onSelect(4);
-         spyOn(component.select, 'emit');
-         expect(component.select.emit).not.toHaveBeenCalled();
-      });
+      expect(component.elementsToShow).toEqual(10);
+      expect(component.indexArray).toEqual(expectedIndexShowing10);
    });
 
-   describe('if initialized when more 6 elements or less', () => {
-      beforeEach(() => {
-         component.options = menuMock1.concat(menuMock2);
-      });
+   it('And user clicks on an element, component emits the element position', () => {
+      spyOn(component.select, 'emit');
+      fixture.detectChanges();
 
-      it('the first element will be displayed followed by 3 dots and concatenated with the last 5 items', () => {
-         expect(component.generateCrumbs()[0]).toEqual(menuMock1[0]);
-         expect(component.generateCrumbs()[1]).toEqual('...');
-         expect(component.generateCrumbs().slice(-4)).toEqual(
-            menuMock2.slice(-4)
-         );
-      });
+      component.onSelect(2);
+      expect(component.select.emit).toHaveBeenCalledWith(2);
+   });
 
-      it('if user clicks on an element displayed after dots, component emits the real position of element', () => {
-         spyOn(component.select, 'emit');
-         component.onSelect(3);
-         expect(component.select.emit).toHaveBeenCalledWith(3);
-      });
+   it('If user clicks on the active element, nothing happens', () => {
+      spyOn(component.select, 'emit');
+      fixture.detectChanges();
+
+      component.onSelect(6);
+      expect(component.select.emit).not.toHaveBeenCalled();
+   });
+
+   it('Should show first element, 3 dots and the last 4 elements ', () => {
+      fixture.detectChanges();
+
+      expect(component.elementsToShow).toEqual(5);
+      expect(component.indexArray).toEqual(expectedIndexShowing5);
+   });
+
+   it('Should emit real position when user click on dots', () => {
+      fixture.detectChanges();
+
+      spyOn(component.select, 'emit');
+      component.onSelect(-1);
+      expect(component.select.emit).toHaveBeenCalledWith(2);
+   });
+
+   it('Should get option when index is greater than -1', () => {
+      fixture.detectChanges();
+      expect(component.getOption(3)).toEqual(menuMock[3]);
+   });
+
+   it('Should get 3 dots when index is -1', () => {
+      fixture.detectChanges();
+      expect(component.getOption(-1)).toEqual('...');
+   });
+
+   it('Should recalculate initPos and elements to show when changes', () => {
+      fixture.detectChanges();
+      expect(component.indexArray).toEqual(expectedIndexShowing5);
+
+      component.ngOnChanges({});
+      expect(component.indexArray).toEqual(expectedIndexShowing5);
+
+      component.ngOnChanges({ elementsToShow: new SimpleChange(5, 7, true) });
+      component.elementsToShow = 7;
+
+      expect(component.indexArray).toEqual([0, 1, 2, 3, 4, 5, 6]);
+
+      component.options = [...menuMock, 'section7'];
+      component.ngOnChanges({ options: new SimpleChange(menuMock, component.options, true)});
+      expect(component.indexArray).toEqual([0, -1, 2, 3, 4, 5, 6, 7]);
    });
 });