+++ src/lib/st-breadcrumbs/st-breadcrumbs.component.ts
@@ -8,38 +8,89 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {
+   ChangeDetectionStrategy,
+   Component,
+   EventEmitter,
+   Input,
+   OnChanges,
+   OnInit,
+   Output,
+   SimpleChanges
+} from '@angular/core';
+import { range as _range } from 'lodash';
 
+/**
+ * @description {Component} [Breadcrumbs]
+ *
+ * The breadcrumb is s a type of secondary navigation
+ * element that shows a trail for allowing users to keep track of their location.
+ *
+ * @example
+ *
+ * {html}
+ *
+ * ```
+ * <st-breadcrumbs
+ *    [options]="options"
+ *    (select)="onSelect($event)"
+ *    qaTag="breadcrumbs">
+ * </st-breadcrumbs>
+ * ```
+ *
+ */
 @Component({
    selector: 'st-breadcrumbs',
    templateUrl: './st-breadcrumbs.component.html',
-   styleUrls: ['./st-breadcrumbs.component.scss']
+   styleUrls: ['./st-breadcrumbs.component.scss'],
+   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class StBreadCrumbsComponent {
+export class StBreadCrumbsComponent implements OnInit, OnChanges {
+   /** @Input {string[]} [options=\[\]] List of navigation parts for show in breadcrumb */
    @Input() options: string[] = [];
+   /** @Input {number} [elementsToShow=5] Max number of elements to show. */
+   @Input() elementsToShow: number = 5;
+   /** @Input {string} [qaTag=''] Id value for qa test */
    @Input() qaTag: string;
+
+   /** @Output {number} [select] Notify any item selection */
    @Output() select: EventEmitter<number> = new EventEmitter<number>();
 
-   generateCrumbs(): string[] {
-      if (this.options.length <= 6) {
-         return [...this.options];
-      } else {
-         return this.options
-            .slice(0, 1)
-            .concat(['...'])
-            .concat(this.options.slice(-4));
+   public indexArray: number[] = [];
+   private _initPos: number = 0;
+
+   public ngOnInit(): void {
+      this.calculateStart(this.options, this.elementsToShow);
+   }
+
+   public ngOnChanges(changes: SimpleChanges): void {
+      if (changes.options || changes.elementsToShow) {
+         const currElsToShow: number = changes.elementsToShow ? changes.elementsToShow.currentValue : this.elementsToShow;
+         const currOptions: string[] = changes.options ? changes.options.currentValue : this.options;
+         this.calculateStart(currOptions, currElsToShow);
       }
    }
 
-   onSelect(index: number): void {
-      if (index + 1 < this.options.length) {
-         if (this.options.length <= 6 || index === 0) {
-            this.select.emit(index);
-         } else {
-            let calculatedIndex: number;
-            calculatedIndex = this.options.length - (6 - index);
-            this.select.emit(index);
-         }
+   public onSelect(index: number): void {
+      const toEmit: number = index > -1 ? index : this._initPos - 1;
+      if (toEmit < (this.options.length - 1)) {
+         this.select.emit(toEmit);
+      }
+   }
+
+   public getOption(index: number): string {
+      return index > -1 ? this.options[index] : '...';
+   }
+
+   private calculateStart(options: string[], elementsToShow: number): void {
+      if (options.length > elementsToShow) {
+         this._initPos = options.length - elementsToShow + 1; // use +1 because first element is fixed
+         this.indexArray = _range(this._initPos, this._initPos + elementsToShow - 1);
+         this.indexArray.unshift(-1); // Add dots fixed element
+         this.indexArray.unshift(0); // Add first fixed element
+      } else {
+         this._initPos = 0;
+         this.indexArray = _range(0, options.length);
       }
    }
 }